Feat/tests merchant republic (#22)

+++ src/ImerchantRepublic.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.9;
 
 interface MerchantRepublicI {
 
-    function guildsVerdict(uint256 proposalId, bool verdict) external;
+    function guildsVerdict(uint48 proposalId, bool verdict) external;
     function getProposalCount() external returns(uint256);
     function setSilverSeason()
         external
